+++ collections/ansible_collections/deamen/certificate/CHANGELOG.rst
@@ -4,6 +4,21 @@ Deamen Certificate Collection Release Notes
 
 .. contents:: Topics
 
+v1.1.0
+======
+
+Release Summary
+---------------
+
+This release adds deploy_certificates module to the collection.
+It is favoured over the deploy_certificate role.
+The role will be deprecated in the next release.
+
+New Modules
+-----------
+
+- deploy_certificate - Deploy certificates and keys to specified locations
+
 v1.0.0
 ======